+++ update.c
@@ -129,7 +129,7 @@ static int ccat_update_release(struct inode *const i, struct file *const f)
  *
  * Return: the number of bytes written, or 0 if EOF reached
  */
-static int ccat_update_read(struct file *const f, char __user * buf, size_t len,
+static long ccat_update_read(struct file *const f, char __user * buf, size_t len,
 			    loff_t * off)
 {
 	struct update_buffer *update = f->private_data;
@@ -158,7 +158,7 @@ static int ccat_update_read(struct file *const f, char __user * buf, size_t len,
  * Return: the number of bytes written, or 0 if flash end is reached
  */
 
-static int ccat_update_write(struct file *const f, const char __user * buf,
+static long ccat_update_write(struct file *const f, const char __user * buf,
 			     size_t len, loff_t * off)
 {
 	struct update_buffer *const update = f->private_data;
